dbg_ioctl_code: change argument to ULONG

+++ driver/lib/dbgcommon.c
@@ -171,7 +171,7 @@ const char *dbg_usbd_status(USBD_STATUS status)
         return "USBD_STATUS_?";
 }
 
-const char *dbg_ioctl_code(int ioctl_code)
+const char *dbg_ioctl_code(ULONG ioctl_code)
 {
 	static_assert(sizeof(ioctl_code) == sizeof(IOCTL_USBIP_VHCI_PLUGIN_HARDWARE), "assert");
